going to add texture loading

+++ MM3DECS/ECS/Engine.swift
@@ -66,8 +66,8 @@ extension Engine : RendererDelegate{
 extension Engine : MTKViewDelegate {
     func mtkView(_ view: MTKView, drawableSizeWillChange size: CGSize) {
         let aspect = Float(view.bounds.width) / Float(view.bounds.height)
-        Renderer.params.width = Float(view.bounds.width) //UInt32(size.width)
-        Renderer.params.height = Float(view.bounds.height)//UInt32(size.height)
+        Renderer.params.width = uint(view.bounds.width) //UInt32(size.width)
+        Renderer.params.height = uint(view.bounds.height)//UInt32(size.height)
         logger.debug("inside Engine: width: \(Renderer.params.width), height: \(Renderer.params.height)")
         sceneManager?.updateCurrentSceneCameraAspectRatio(with: aspect)
     }
